modules/beta/emrstudio-on-eks/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
attrs==23.2.0
    # via
    #   cattrs
    #   jsii
aws-cdk-asset-awscli-v1==2.2.202
    # via aws-cdk-lib
aws-cdk-asset-kubectl-v20==2.1.2
    # via aws-cdk-lib
aws-cdk-asset-node-proxy-agent-v6==2.0.3
    # via aws-cdk-lib
aws-cdk-cloud-assembly-schema==36.0.24
    # via aws-cdk-lib
aws-cdk-lambda-layer-kubectl-v29==2.0.0
    # via -r requirements.in
aws-cdk-lib==2.155.0
    # via
    #   -r requirements.in
    #   aws-cdk-lambda-layer-kubectl-v29
    #   cdk-nag
boto3==1.35.11
    # via -r requirements.in
botocore==1.35.11
    # via
    #   boto3
    #   s3transfer
cattrs==23.2.3
    # via jsii
cdk-nag==2.28.195
    # via -r requirements.in
cffi==1.16.0
    # via cryptography
constructs==10.3.0
    # via
    #   -r requirements.in
    #   aws-cdk-lambda-layer-kubectl-v29
    #   aws-cdk-lib
    #   cdk-nag
cryptography==43.0.1
    # via
    #   -r requirements.in
    #   pyopenssl
importlib-resources==6.4.0
    # via jsii
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
jsii==1.103.1
    # via
    #   aws-cdk-asset-awscli-v1
    #   aws-cdk-asset-kubectl-v20
    #   aws-cdk-asset-node-proxy-agent-v6
    #   aws-cdk-cloud-assembly-schema
    #   aws-cdk-lambda-layer-kubectl-v29
    #   aws-cdk-lib
    #   cdk-nag
    #   constructs
publication==0.0.3
    # via
    #   aws-cdk-asset-awscli-v1
    #   aws-cdk-asset-kubectl-v20
    #   aws-cdk-asset-node-proxy-agent-v6
    #   aws-cdk-cloud-assembly-schema
    #   aws-cdk-lambda-layer-kubectl-v29
    #   aws-cdk-lib
    #   cdk-nag
    #   constructs
    #   jsii
pycparser==2.22
    # via cffi
pyopenssl==24.2.1
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via
    #   botocore
    #   jsii
s3transfer==0.10.2
    # via boto3
six==1.16.0
    # via python-dateutil
typeguard==2.13.3
    # via
    #   aws-cdk-asset-awscli-v1
    #   aws-cdk-asset-kubectl-v20
    #   aws-cdk-asset-node-proxy-agent-v6
    #   aws-cdk-cloud-assembly-schema
    #   aws-cdk-lambda-layer-kubectl-v29
    #   aws-cdk-lib
    #   cdk-nag
    #   constructs
    #   jsii
typing-extensions==4.12.2
    # via jsii
urllib3==1.26.19
    # via
    #   -r requirements.in
    #   botocore
